store counter of restarts in NVS (Non-Volatile Storage)
In preparation of #748, this counter will serve as part of an "unique"
token in the log file name ; ensuring that after a reboot we will
not overwrite an existing log file.

The counter is stored in NVS, and is incremented at each class instanciation;
which should roughly translate to every reboot.
(Don't know about the deepsleep for the moment).

The counter will reset to 1 after 99.999.999 restarts, which should gives us
plenty of time.

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>